EggNOG database annotations

EggNOG is a database of biological information hosted by the EMBL. It is based on the original idea of COGs (clusters of orthologous groups) and expands that idea to non-supervised orthologous groups constructed from numerous organisms. eggNOG stands for evolutionary genealogy of genes: Non-supervised Orthologous Groups.

EggNOG()

Value

EggNOG.

Note

Updated 2020-07-23.

Examples

options(acid.test = TRUE)
x <- EggNOG()

#> → Importing eunog.tsv.gz at https://tests.acidgenomics.com/basejump/v0.13 using vroom::`vroom()`.

#> → Importing nog.tsv.gz at https://tests.acidgenomics.com/basejump/v0.13 using vroom::`vroom()`.

print(x)

#> EggNOG 0.1.2 of length 2
#> ids(10): COG5153 COG5156 ... ENOG410KJ10 ENOG410KJ1P
#> categories(18): Amino acid transport and metabolism Carbohydrate transport and metabolism ... Signal transduction mechanisms Transcription